src/fyodorov_utils/requirements.txt
fastapi
asyncio
PyJWT
starlette
supabase
pydantic>=2.11.3
fyodorov_llm_agents==0.4.65